[AArch64] Lower "llvm.ptrauth" signed global wrappers.
This is a large patch, and basically adds end-to-end support
for generating authenticated global relocation.

At the IR level, this currently uses a flawed representation
(because of downstream compatibility concerns): instead of a
"ptrauth sign" ConstantExpr, a signed global pointer is represented
using a global variable wrapper, in the special section "llvm.ptrauth".

This has some ramifications at the IR level, but is mostly
straightforward.

In data, signed global references are lowered to the arm64e
AUTHENTICATED_POINTER relocation, via AArch64AuthMCExpr.

In code, signed global pointers can be materialized using:
- a code sequence, via the MOVaddrPAC (hardened) pseudo
- a compiler GOT section, __DATA,__auth_ptr.
At the ISel level, the reference is represented using a
"PtrAuthGlobalAddress" node.