[InstCombine] Combine ptrauth intrinsics into call operand bundles.
Since the "ptrauth" call operand bundles really describe a folded
"llvm.ptrauth.auth" operation, there are optimizations we can
make to elide redundant sign/auth pairs.

Note that this needs to be careful about preserving the integrity
of the intermediate unauthenticated values, as that's the sole
purpose of having operand bundles in the first place.