[AArch64] Support @AUTH symbol variant and MachO reloc.
This adds support for a new symbol variant, "@AUTH".  This describes a
symbol reference that needs to be signed at load-time:
  .quad _sym@AUTH(ib,123[,addr])
  .quad (_sym + 5)@AUTH(ib,123[,addr])
  .quad (_sym - 5)@AUTH(ib,123[,addr])

At the asm level, it requires parameters, which makes it slightly more
complicated than existing variants: because of that, it's implemented
separately, in an early parsePrimaryExpr hook.

When emitting MachO, it maps to the ARM64_RELOC_AUTHENTICATED_POINTER
relocation.  It's not currently supported in other object writers.